requirements.txt
# =================================================================
# REQUIREMENTS.TXT - Dependencias del proyecto
# Jorge Aguirre Flores - Maquillaje Permanente
# =================================================================

# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Templates HTML
jinja2==3.1.2

# HTTP Requests (Meta CAPI)
requests==2.31.0

# Formularios/Uploads
python-multipart==0.0.6

# Procesamiento de Imágenes (WebP)
pillow==10.2.0

# PostgreSQL (Supabase) - Persistencia de Atribución
psycopg2-binary==2.9.9

# Validación de datos (Modelos Pydantic)
pydantic==2.5.0
